+++ wsi_service/custom_models/service_status.py
@@ -8,9 +8,7 @@ from wsi_service.models.commons import ServiceStatus
 class PluginInfo(BaseModel):
     name: str
     version: str
-    supported_file_extensions: List[str]
 
 
 class WSIServiceStatus(ServiceStatus):
     plugins: List[PluginInfo]
-    plugins_default: dict