+++ wsi_service_base_plugins/openslide/wsi_service_plugin_openslide/__init__.py
@@ -1,16 +1,15 @@
+import glob
+import os
+import pathlib
+
 from wsi_service_plugin_openslide.slide import Slide
 
-supported_file_extensions = [
-    ".bif",
-    ".mrxs",
-    ".ndpi",
-    ".scn",
-    ".svs",
-    ".tiff",
-    ".tif",
-    ".vsf",
-    "vsf-folder",
-]
+
+def is_supported(filepath):
+    if os.path.isfile(filepath):
+        return pathlib.Path(filepath).suffix in [".bif", ".mrxs", ".ndpi", ".scn", ".svs", ".tiff", ".tif"]
+    else:
+        return len(glob.glob(os.path.join(filepath, "*.vsf"))) > 0
 
 
 async def open(filepath):