+++ wsi_service_base_plugins/openslide/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.4.0
+
+- updated for new plugin concept
+
 ## 1.3.1
 
 - Updated dependencies