+++ wsi_service_base_plugins/wsidicom/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.4.0
+
+- updated for new plugin concept
+
 ## 0.3.1
 
 - updated wsidicom